Extract reusable server GC hook registry
WebRTC support will also want to be able to clean up any records leaked by
other servers, so extract the periodic cleanup into something reusable.
